AFIRGEN FINAL/main backend/requirements.txt
chromadb==1.0.20
fastapi==0.116.1
httpx==0.28.1
mysql_connector_repackaged==0.3.1
pydantic==2.11.7
python_magic==0.4.27
uvicorn==0.35.0
llama-cpp-python==0.3.1  
openai-whisper==20240930
transformers==4.52.4
torch==2.7.1
filelock==3.16.1
Pillow==11.3.0
ffmpeg-python==0.2.0
boto3==1.35.94
aws-xray-sdk==2.14.0

# Backend Optimization Dependencies
redis==5.2.1
celery==5.4.0
hypothesis==6.122.3
prometheus-client==0.21.1
structlog==24.4.0
psutil==6.1.1

# OpenTelemetry Dependencies
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-httpx==0.50b0
opentelemetry-instrumentation-redis==0.50b0
opentelemetry-instrumentation-mysql==0.50b0
opentelemetry-instrumentation-celery==0.50b0
opentelemetry-exporter-otlp==1.29.0

# Performance Testing Dependencies
pytest-benchmark==5.1.0
locust==2.33.0
